DP-631 Merge in batch job from s3-dataset-scanner
Merge in the dataset scanning batch job from the s3-dataset-scanner
repository, turning this repository into a general collection of batch
jobs for the dataplatform.

This commit brings an (almost) minimal set of changes, merging only the
code that was already merged into master in the s3-dataset-scanner
repository. Follow-up commits will address issues that were raised in
https://github.oslo.kommune.no/origo-dataplatform/s3-dataset-scanner/pull/1
after it was merged, and rewrite the other pending PRs for this
repository instead.

The name of the repository will be changed in Github too once the
proposed name is settled.